+++ codeforces_core/interfaces/Conf.py
@@ -1,5 +1,0 @@
-from typing import Protocol
-
-
-class ConfInterface(Protocol):
-  prog_id: str