+++ tests/e2e/test_submit.py
@@ -1,4 +1,14 @@
+import asyncio
+import os
+
+# test
+from . import test_e2e_dir
+
+E2E_CONFIG_FILE = os.path.join(test_e2e_dir, 'e2e_test_config.py')
+
+pytest_plugins = ('pytest_asyncio', )
+
+
 def test_submit():
-  assert False  # TODO
-  from codeforces_core.submit import submit
-  submit(fhh, cid='1234', level='D2', filename='Main.cpp', lang_id='9876')
+  # check comments in submit
+  pass