+++ tests/unit/mock/main.cpp
@@ -1,0 +1,5 @@
+#include <bits/stdc++.h>
+int main() {
+  printf("demo main\n");
+  return 0;
+}