+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add auto doc
 - Add auto test(e2e, unit, coverage report)
 - Add vscode-docker container support
+- Reconstruct submit, separate the function as submit / websocket listening / polling submission page
 
 ### Features